PerfAPI: One more warning for long long unsigned int
svn:sync-xref-src-repo-rev: r34661

+++ src/VBox/Main/linux/PerformanceLinux.cpp
@@ -1,4 +1,4 @@
-/* $Id: PerformanceLinux.cpp 11385 2008-08-13 12:26:24Z aleksey.ilyushin@oracle.com $ */
+/* $Id: PerformanceLinux.cpp 11386 2008-08-13 12:32:33Z aleksey.ilyushin@oracle.com $ */
 
 /** @file
  *
@@ -145,7 +145,7 @@ int CollectorLinux::getRawProcessStats(RTPROCESS process, uint64_t *cpuUser, uin
     pid_t pid2;
     char c;
     int iTmp;
-    uint64_t u64Tmp;
+    long long unsigned int u64Tmp;
     unsigned uTmp;
     unsigned long ulTmp;
     ULONG u32user, u32kernel;
